Rename logging.py to logging_config.py to avoid stdlib name collision
