+++ mosaico-sdk-py/src/mosaicolabs/handlers/topic_writer.py
@@ -20,7 +20,7 @@ from ..helpers import pack_topic_resource_name
 from ..comm.do_action import _do_action
 from ..enum import FlightAction, OnErrorPolicy
 from .config import WriterConfig
-from ..logging import get_logger
+from ..logging_config import get_logger
 
 # Set the hierarchical logger
 logger = get_logger(__name__)
